feat: add multiplatform image processing utility
This commit introduces an ImageProcessor utility
to handle image manipulation, such as resizing,
in a cross-platform manner.

[Desc]
 -- Added an ImageProcessor interface in the :core module to define the contract for image processing.
-- Implemented the expect/actual pattern in the :data module for platform-specific logic.
-- Provided the actual implementation for Android using the android.graphics.Bitmap factory.
-- Provided the actual implementation for iOS using UIKit to handle image resizing.
-- Included the generated Room schema file (2.json) corresponding to the recent database version bump.

+++ core/src/commonMain/kotlin/com/bitlabbr/minhadespensa/core/domain/model/CatalogProduct.kt
@@ -32,7 +32,7 @@ data class CatalogProduct(
     val name: String,
     val brand: String? = null,
     val measureUnit: MeasureUnit,
-    val netWeight: Long,
+    val netWeight: Double,
     val thumbnailUrl: String? = null,
     val updatedAt: Long,
     val isDeleted: Boolean = false,
